Workaround for systems w/o TCSASOFT tcsetattr() flag
This flag is used on *BSD when calling tcsetattr() to prevent it from
changing speed, duplex, and parity.  GNU says we should use the CIGNORE
flag to c_cflag, but that doesn't exist so ... we rely on our initial
tcgetattr() and prey that nothing changes on the TTY before we exit and
restore with tcsetattr().

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>