Use POSIX st_mtim over intermediate st_mtimespec
For sub second resolution in stat() et al POSIX added a struct timespec,
with down to nano second resolution.  There was a couple of intermediate
names used before the standard settled.  For mtime it was st_mtimespec,
which OpenBSD and a few others now have conventient aliases for.  GNU
however does not have this and the correct way, instead of adding a
define in def.h, is to update the sources.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>