Use POSIX getline() instead of *BSD fgetln()
This patch converts mg to use the POSIX getline() API instead of the
*BSD specific fgetln().  The latter is rather elegant, but getline() is
not too shabby either, the only extra trouble is that you must free the
line after the while()-loop -- getline() allocates and reallocates,
i.e. reuses the currently allocated line, just like fgetln() uses the
internal buffers of the C library.

The biggest payoff with this patch is that getline() works on a lot more
operating systems, including OpenBSD so no extra porting effort is
needed.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>