Use configure's PACKAGE_STRING as Mg version, v3.0
The Mg version has been the same since 1986, time to up it to a more
relevant version.  Since Mg 2a is the origin of the OpenBSD version and
there exist an Mg 3a <http://www.bengtl.net/files/mg3a/>, I chose to go
with 3.0, since it lexically is somewhere in between. Sort of.

The version is seen both in the tarball name, the configure script, and
M-x emacs-version inside Mg.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>