Restore support for theo analyze etc. on OpenBSD
This patch restores support for theo.c by including it in the build by
default on OpenBSD systems.

We are now level with the upstream OpenBSD mg again.  The code should
now be functionally equivalent and result in (almost) the same binary
as the original code.  Except for the new version string.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>