Undefine _FORTIFY_SOURCE to prevent annoying GCC warnings
There are quite a few C library calls that are marked to warn if the
return value goes unchecked.  For instance the write() in region.c,
function pipeio(), where there is no point in checking the return value.
We're in a "limp home" mode since execv() just failed and the call after
the write() is _exit(1), so we really do not care.  There are quite a
few constructs like that.  Previously you could (void) mark such calls,
but now you need to check them or do what this patch does.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>