configure: Check for -ltermcap and term.h
This patch adds clever detection for a -ltermcap capable library, which
can be any of: libtermcap, libtinfo, libcurses, or libncurses.  It also
checks for term.h, not termcap.h, because we like cookies.

Configure is clever in that it adds the found library to $LIBS, which is
what mg is later linked with, along with $(lite_LIBS) which pkg-config
provides us with.  So, this patch also remove -ltermcap from mg_LDADD,
since anything that has tgoto() will do.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>