Keep current buffer if switch-to-buffer is aborted
For some reason I find myself quite often in the situation that I want
to C-x b to a different buffer but on entering the buffer name I
decide against it and C-g abort it.

mg(1) then switches me to *scratch*. IMO that is the least useful
thing mg(1) can do. This brings it in line with what emacs does and
keeps the current buffer.

Signed-off-by: Florian Obser <florian@openbsd.org>