Present the default choice before the colon in prompts
GNU Emacs always, or very nearly always, prints the default
choice before the colon in any given prompt.  For example,
you'll see

        "Kill buffer (default foo): "

not

        "Kill buffer: (default foo) "

This patch moves all default choices before the colon in all
of mg's prompts.

While here, add missing spaces in two prompts in tags.c and
capitalize the tables prompt in tags.c.  The first word of
a minibuffer prompt is typically capitalized, and all others
in mg(1) are currently capitalized except for "apropos",
which is exempt because in that case it's meant to evoke the
utility of the same name.

Signed-off-by: Scott Cheloha <scottcheloha@gmail.com>