Backups saved to wrong directory when running as another user
When run as a different effective user, i.e. when invoked via su and
backup-to-home-directory is enabled, mg tries to write backups to the
wrong directory.

Problem pointed out and diff provied by Lucas Gabriel Vuotto
<lvuotto92 () gmail ! com>, thanks!

Subsequently slacked on for nearly a year by yours truly.

Then remembered when Han Boetes <hboetes () gmail ! com> came up with
a similar diff because of a problem report by Mark Willson where it
turned out that getlogin(2) is not very portable.

Signed-off-by: Florian Obser <florian@openbsd.org>