debian/rules: Simplify, remove ancient README and license file
This project uses GNU configure & build, so we don't have to jump
through hoops anymore to build.  Standard debhelper build is OK.

However, we still need to prune some files:
- README, from upstream OpenBSD, not needed
- UNLICENSE, licensing info is in the canonical copyright file

Also, debhelper needs to be told our non-standard ChangeLog.md so it can
compress it, and not replace it with the debian/changelog.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>