Amend showcpos() to show more information about how mg is working
The additions help while developing mg by not having to go into gdb to
see somethings mg believes about itself. The extra data shown could be
removed in future, but it is benign so perhaps it could stay.

Also, do not fake a '\n' on the end of buffer. It is confusing.  However
do not change the lie about '\n' being on the end of every line within
the buffer. Hopefully, one day, that will no longer be a lie.

Signed-off-by: Mark Lumsden <lum@openbsd.org>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>