Manually revert d165803: blink-and-insert can be called from ~/.mg
I totally forgot about the instance where blink-and-insert can be called
from the mg startup file.  My previous diff broke that instance.  So
this diff reverts my change but adds a man page line to blink-and-insert
in case someone should try to use it interactively, again.  Also, should
anyone ever want to add an interactive version of blink-and-insert in
the future, the code is there.  And the code could be useful in another,
related instance.

Signed-off-by: Mark Lumsden <lum@openbsd.org>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>