Allow fns with >0 param receive and process when eval from startup file
This diff does treat the '(' and ')' chars differently during evaluation
than previously, in-so-far as they are not ignored if they are at the
end or start of a line now. However, even though these characters are
not ignored, this diff should not change the behaviour of an extant .mg
file, with '(' and ')' chars at the end and start of a line.  This
situation is accomodated for in this diff (with limited testing though).

Signed-off-by: Mark Lumsden <lum@openbsd.org>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>