log: Minor, use void * intead of KEYMAP
To be able to include log.h in more parts of mg, we can either add
include guards to all files, or patch up the include files we have.

This is a workaround to use void ptr as external (internal) API to
mglog().  Almost all C files already include def.h, but not all
should need to include kbd.h

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>