Fix regression; status line cleared when searching for file to open
While trying to open a file with C-x C-f and tabbing around in the file
system, the status line that displays the current path is cleared.  It
turns out this is a regression from 6def8a9.

This patch reverts part of 6def8a9 and implements a more correct fix to
the problem of lingering status message at switch/kill by name.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>