Allow more than one block of code to exist on a single line.
Allow more than one block of code to exist on a single line. Also, move
the code which expands variables to be ran when variables are discovered
instead of in multiarg() just before execution. This means a variable
who's value is included in anothers', won't change the others value if
its own is changed.

I have also included code, which is commented out at the moment, which
implements a function map specific to interpreter. Not sure if I can
ultimately avoid using it though.

Signed-off-by: Mark Lumsden <lum@openbsd.org>
Signed-off-by: Joachim Wiberg <troglobit@gmail.com>