Drop trailing whitespace on RET (c-mode)
mg(1) has this tendency to leave a lot of trailing whitespaces around
in auto-indent-mode and c-mode which I find annoying.  Yes, there's
delete-trailing-space but it works only on the current line (and not
the whole buffer as in emacs) and one has to manually call it anyway.
Emacs, and even vi in base, are clever in this regard: trailing
whitespaces before the cursor are delete upon RET before adding a new
line.

So, here's the same heuristic for mg when auto-indent-mode or c-mode is
enabled.  It's still possible to leave trailing whitespace in a buffer
in those modes, it only gets a little bit harder to do.

Signed-off-by: Omar Polo <op@openbsd.org>
Signed-off-by: Joachim Wiberg <troglobit@gmail.com>