Change tagvisit (aka visit-tag-table) to immediately load tag file
Before it used to only record the path to the tag file which was loaded
on demand upon find-tag (M-.).  tagvisit did to do a stat + access dance
to know if the path was pointing to a regular, readable file and
loadtags (called lazily by find-tag) trusted those checks...

Instead, just load the tags in tagvisit and drop the lazy mechanics.

Signed-off-by: Omar Polo <op@openbsd.org>