Redcuce *quick* size and make help text in status area permanent
In an effort to improve on the usability for beginners, without getting
in the way of experienced Mg users, the initial help text in the status
area is now shown permanently.  I.e., whenever Mg clears the statis area
we show the help text.

Since we now show the help text in the status area we can drop it from
the *quick* help buffer and reduce the height to save screen estate.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>